feat: Improve study search functionality and remove search modal
