+++ apis/service/study.ts
@@ -16,8 +16,8 @@ export const StudyService = () => {
    * 스터디 검색 - 입력한 키워드가 제목에 포함된 스터디 목록을 반환
    * @api-doc: http://ec2-3-38-27-246.ap-northeast-2.compute.amazonaws.com/swagger-ui/index.html#/03.%20%EC%8A%A4%ED%84%B0%EB%94%94/searchStudy
    */
-  const search = async (keyword?: number, pageNumber?: number, pageSize?: number) => {
-    const req = (await API.get(url, { params: { keyword, pageNumber, pageSize } })) as StudyResponse.Search;
+  const search = async (keyword?: string, pageNumber?: number, pageSize?: number) => {
+    const req = (await API.get(`${url}/search`, { params: { keyword, pageNumber, pageSize } })) as StudyResponse.Search;
     return req.data;
   };
