+++ app/(tabs)/(home)/_layout.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import { Stack } from 'expo-router';
+import { RecommendationProvider } from '@/context/Recommendation';
 
 function MoreStudiesLayout() {
-  return <Stack screenOptions={{ headerShown: false }} initialRouteName="(explore)" />;
+  return (
+    <RecommendationProvider>
+      <Stack screenOptions={{ headerShown: false }} initialRouteName="(explore)" />
+    </RecommendationProvider>
+  );
 }
 
 export default MoreStudiesLayout;
